refactor: remove prev decoder
