+++ crates/binary-cookies/src/decode/meta.rs
@@ -1,5 +1,3 @@
-use std::io::Read;
-
 use oval::Buffer;
 use winnow::{
     error::{ErrMode, Needed},
@@ -18,7 +16,13 @@ pub type MetaOffset = u64;
 #[derive(Clone)]
 #[derive(Debug)]
 pub struct MetaFsm {
-    buffer: Buffer,
+    pub(crate) buffer: Buffer,
+}
+
+impl Default for MetaFsm {
+    fn default() -> Self {
+        Self::new()
+    }
 }
 
 impl MetaFsm {
@@ -32,7 +36,7 @@ impl MetaFsm {
 
     pub fn process(mut self) -> Result<DecodeResult<Self, (Checksum, Option<Metadata>)>> {
         let mut input: StreamIn = StreamIn::new(self.buffer.data());
-        let e = match BinaryCookies::parse_tail.parse_next(&mut input) {
+        let e = match BinaryCookies::decode_tail.parse_next(&mut input) {
             Ok((checksum, meta)) => {
                 return Ok(DecodeResult::Done((checksum, meta)));
             },
@@ -55,39 +59,6 @@ impl MetaFsm {
                 }
                 Ok(DecodeResult::Continue(self))
             },
-        }
-    }
-}
-
-#[derive(Clone, Copy)]
-#[derive(Debug)]
-#[derive(Default)]
-#[derive(PartialEq, Eq, PartialOrd, Ord)]
-pub struct MetaDecoder<R: Read> {
-    pub(crate) rd: R,
-}
-
-impl Default for MetaFsm {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl<R: Read> MetaDecoder<R> {
-    pub fn decode(&mut self) -> Result<(Checksum, Option<Metadata>)> {
-        let mut fsm = MetaFsm::new();
-        loop {
-            self.rd
-                .read_exact(fsm.buffer.space())?;
-            let count = fsm.buffer.available_space();
-            fsm.buffer.fill(count);
-            match fsm.process()? {
-                DecodeResult::Continue(fsm_) => {
-                    fsm = fsm_;
-                    continue;
-                },
-                DecodeResult::Done(done) => return Ok(done),
-            }
         }
     }
 }
