+++ crates/binary-cookies/src/decode/pages.rs
@@ -1,18 +1,12 @@
-use std::io::Read;
-
 use oval::Buffer;
 use winnow::{
     error::{ErrMode, Needed},
     Parser,
 };
 
-use super::{
-    cookies::{CookieHandle, CookiesOffset},
-    DecodeResult, OffsetSize,
-};
+use super::{DecodeResult, OffsetSize};
 use crate::{
     cookie::Page,
-    cursor::CookieCursor,
     decode::StreamIn,
     error::{ParseError, Result},
 };
@@ -48,7 +42,7 @@ impl PagesOffset {
 #[derive(Clone)]
 #[derive(Debug)]
 pub struct PageFsm {
-    buffer: Buffer,
+    pub(crate) buffer: Buffer,
 }
 
 impl PageFsm {
@@ -68,7 +62,7 @@ impl PageFsm {
 
     pub fn process(mut self) -> Result<DecodeResult<Self, Vec<u32>>> {
         let mut input: StreamIn = StreamIn::new(self.buffer.data());
-        let e = match Page::parse_head.parse_next(&mut input) {
+        let e = match Page::decode_head.parse_next(&mut input) {
             Ok(o) => return Ok(DecodeResult::Done(o)),
             Err(e) => e,
         };
@@ -98,75 +92,5 @@ impl PageFsm {
 impl Default for PageFsm {
     fn default() -> Self {
         Self::new()
-    }
-}
-
-#[derive(Clone)]
-#[derive(Debug)]
-pub struct PagesHandle<'a, F: CookieCursor> {
-    pub(crate) file: &'a F,
-    pub(crate) offset: PagesOffset,
-}
-
-impl<'a, F: CookieCursor> PagesHandle<'a, F> {
-    pub fn decoders(
-        &self,
-    ) -> impl Iterator<Item = PageDecoder<'_, <F as CookieCursor>::Cursor<'_>, F>> {
-        self.offset
-            .offset_sizes
-            .iter()
-            .map(|&OffsetSize { offset, size }| PageDecoder {
-                file: self.file,
-                rd: self.file.cursor_at(offset),
-                offset,
-                size,
-            })
-    }
-
-    pub fn into_decoders(
-        self,
-    ) -> impl Iterator<Item = PageDecoder<'a, <F as CookieCursor>::Cursor<'a>, F>> {
-        self.offset
-            .offset_sizes
-            .into_iter()
-            .map(|OffsetSize { offset, size }| PageDecoder {
-                file: self.file,
-                rd: self.file.cursor_at(offset),
-                offset,
-                size,
-            })
-    }
-}
-
-#[derive(Clone, Copy)]
-#[derive(Debug)]
-#[derive(PartialEq, Eq, PartialOrd, Ord)]
-pub struct PageDecoder<'a, R: Read, F: CookieCursor> {
-    file: &'a F,
-    rd: R,
-    offset: u64,
-    size: u32,
-}
-
-impl<'a, R: Read, F: CookieCursor> PageDecoder<'a, R, F> {
-    pub fn decode(&mut self) -> Result<CookieHandle<'a, F>> {
-        let mut fsm = PageFsm::new();
-        loop {
-            self.rd
-                .read_exact(fsm.buffer.space())?;
-            let count = fsm.buffer.available_space();
-            fsm.buffer.fill(count);
-            match fsm.process()? {
-                DecodeResult::Continue(fsm_) => {
-                    fsm = fsm_;
-                    continue;
-                },
-                DecodeResult::Done(cookie_offset_in_page) => {
-                    let cookies_offset =
-                        CookiesOffset::new(self.offset, self.size, &cookie_offset_in_page);
-                    return Ok(CookieHandle::new(self.file, cookies_offset));
-                },
-            }
-        }
     }
 }
